LayoutTests/fast/viewport/viewport-84-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument value "" for key "width" not recognized. Content ignored.
ALERT: viewport size 64x480 scale 5 with limits [5, 5] and userScalable true
